Refa: import issue. (#9958)
### What problem does this PR solve?


### Type of change

- [x] Refactoring
